refactor: 도메인 중심 구조 패키지 수정 완료
- user/domain, user/application, user/adapter 패키지 수정
- emotion/domain, emotion/application, emotion/adapter 패키지 수정
- common/model 패키지 수정
- 모든 import 경로 업데이트
