+++ src/main/java/com/melog/melog/emotion/application/EmotionKeywordPersistencePort.java
@@ -1,7 +1,7 @@
-package com.melog.melog.application.port.out;
+package com.melog.melog.emotion.application;
 
-import com.melog.melog.domain.emotion.EmotionKeyword;
-import com.melog.melog.domain.emotion.EmotionRecord;
+import com.melog.melog.emotion.domain.EmotionKeyword;
+import com.melog.melog.emotion.domain.EmotionRecord;
 
 import java.util.List;
 import java.util.Optional;
