+++ src/main/java/com/melog/melog/emotion/application/EmotionRecordService.java
@@ -1,11 +1,10 @@
-package com.melog.melog.application.service;
+package com.melog.melog.emotion.application;
 
-import com.melog.melog.application.port.in.EmotionRecordUseCase;
-import com.melog.melog.application.port.out.*;
-import com.melog.melog.domain.emotion.*;
-import com.melog.melog.domain.model.request.*;
-import com.melog.melog.domain.model.response.*;
-import com.melog.melog.domain.user.User;
+import com.melog.melog.common.model.request.*;
+import com.melog.melog.common.model.response.*;
+import com.melog.melog.emotion.domain.*;
+import com.melog.melog.user.application.UserPersistencePort;
+import com.melog.melog.user.domain.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
